pyproject.toml
[tool.poetry]
name = "manga_tools"
version = "0.1.0"
description = ""
authors = ["Gioo9le <g.minoccari@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
requests = "^2.27.1"
bs4 = "^0.0.1"
opencv-python = "^4.5.5"
Pillow = "^9.0.1"
PyPDF2 = "^1.26.0"
uvicorn = "^0.17.6"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
